+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+### 1.0.2
+
+1. Add 'Clear archived tabs' command;
+2. The detail of each archived tab now uses relative path if available.
+3. Add stored archived tabs support.
+
 ### 1.0.1
 
 Add 'Close all diff tabs' support